+++ src/plugins/replication/replication-helper.ts
@@ -1,0 +1,2 @@
+// does nothing
+export const DEFAULT_MODIFIER = (d: any) => Promise.resolve(d);